Change how validatePlayers() works
On validatePlayers(), reject if the opponent's presence is equal to
offline instead of reject when the opponent presence is not equal to
online, because the bot should be able to challenge the opponent, even
though the opponent's presence is equal to idle or dnd.

Set game to null if the players are not valid.