Add a way to more players play at the same time
In bot.js create the object matches and pass it as parameter to a
command when it is called.

Change every command related to game functions in order to accept more
than two players playing at the same time.

Add matchMethods.js module to handle match related actions.

Add unit tests for createMatch, deleteMatch, and findMatch.